+++ src/main/java/com/cstav/genshinstrument/networking/packet/instrument/InstrumentPacket.java
@@ -21,12 +21,12 @@ public class InstrumentPacket implements ModPacket {
 
     private final NoteSound sound;
     private final InteractionHand hand;
-    private final float pitch;
+    private final int pitch;
 
     private final ResourceLocation instrumentId;
     private final NoteButtonIdentifier noteIdentifier;
 
-    public InstrumentPacket(NoteSound sound, float pitch, InteractionHand hand,
+    public InstrumentPacket(NoteSound sound, int pitch, InteractionHand hand,
             ResourceLocation instrumentId, NoteButtonIdentifier noteIdentifier) {
         this.sound = sound;
         this.hand = hand;
@@ -38,7 +38,7 @@ public class InstrumentPacket implements ModPacket {
     public InstrumentPacket(FriendlyByteBuf buf) {
         sound = NoteSound.readFromNetwork(buf);
         hand = buf.readEnum(InteractionHand.class);
-        pitch = buf.readFloat();
+        pitch = buf.readInt();
 
         instrumentId = buf.readResourceLocation();
         noteIdentifier = NoteButtonIdentifier.readIdentifier(buf);
